KNOX-1664 - Miscellaneous code cleanup
* Unnecessary boxing/unboxing
* Collection.toArray()
* String concatenation in loop
* StringBuffer -> StringBuilder where possible
* Identical catch block collapse
* Inline redundant variable
* Collections methods (addAll, isEmpty)
* Use isEmpty where applicable
* Remove redundant initializer

+++ gateway-server/src/main/java/org/apache/knox/gateway/filter/PortMappingHelperHandler.java
@@ -64,7 +64,7 @@ public class PortMappingHelperHandler extends HandlerWrapper {
 
       for(final Map.Entry<String, Integer> entry: config.getGatewayPortMappings().entrySet()) {
 
-        if(entry.getValue().intValue() == config.getGatewayPort()) {
+        if(entry.getValue() == config.getGatewayPort()) {
           defaultTopologyRedirectContext = "/" + config.getGatewayPath() + "/" + entry.getKey();
           break;
         }
